Refactor shared styled containers
