+++ packages/lib-classifier/src/components/Classifier/components/Layout/components/NoMaxWidth/NoMaxWidth.js
@@ -3,12 +3,19 @@ import { Box, Grid } from 'grommet'
 
 import Banners from '@components/Classifier/components/Banners'
 import FeedbackModal from '@components/Classifier/components/Feedback'
-import ImageToolbar from '@components/Classifier/components/ImageToolbar'
 import MetaTools from '@components/Classifier/components/MetaTools'
 import QuickTalk from '@components/Classifier/components/QuickTalk'
 import SubjectViewer from '@components/Classifier/components/SubjectViewer'
 import TaskArea from '@components/Classifier/components/TaskArea'
 import FieldGuide from '@components/Classifier/components/FieldGuide'
+import {
+  ViewerGrid,
+  StyledSubjectContainer,
+  StyledImageToolbarContainer,
+  StyledImageToolbar,
+  StyledTaskAreaContainer,
+  StyledTaskArea
+} from '../shared/StyledContainers'
 
 const ContainerGrid = styled(Grid)`
   position: relative;
@@ -44,59 +51,6 @@ const ContainerGrid = styled(Grid)`
     grid-template-rows: auto auto;
     margin: 0;
   }
-`
-
-export const ViewerGrid = styled(Grid)`
-  ${props => props.hasSurveyTask
-    ? css`
-        @media screen and (min-width: 70rem) {
-          position: sticky;
-          top: 10px;
-        }
-      `
-    : css`
-        @media screen and (min-width: 769px) {
-          position: sticky;
-          top: 10px;
-        }
-      `}
-  height: fit-content;
-  grid-area: viewer;
-  grid-template-columns: auto clamp(3rem, 10%, 4.5rem);
-  grid-template-rows: auto;
-  grid-template-areas: 'subject toolbar';
-`
-
-const StyledSubjectContainer = styled(Box)`
-  grid-area: subject;
-  position: sticky;
-`
-
-const StyledImageToolbarContainer = styled.div`
-  grid-area: toolbar;
-`
-
-const StyledImageToolbar = styled(ImageToolbar)`
-  position: sticky;
-  top: 10px;
-`
-
-const StyledTaskAreaContainer = styled.div`
-  grid-area: task;
-`
-
-const StyledTaskArea = styled(Box)`
-  ${props => props.hasSurveyTask ? css`
-    @media screen and (min-width: 70rem) {
-      position: sticky;
-      top: 10px;
-    }
-  ` : css`
-    @media screen and (min-width: 769px) {
-      position: sticky;
-      top: 10px;
-    }
-  `}
 `
 
 export default function NoMaxWidth({
